refactor(mobile): extract queue provider re-grade engine
